backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Visions – Interior Design Ideator API"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "google-genai>=1.65.0",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.10.0",
    "python-multipart>=0.0.22",
    "sqlmodel>=0.0.37",
    "supabase>=2.0.0",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "pyrefly>=0.54.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.4",
]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4", "PTH", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["app"]
omit = ["tests/*"]
